requirements.txt
# Core Dependencies
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Machine Learning & Vector Operations
sentence-transformers>=2.2.0
numpy>=1.24.0

# Development & Testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
python-dotenv
sentence-transformers
scikit-learn
tenacity
